% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmarFunctions.R
\name{addMargin}
\alias{addMargin}
\title{Adds a margin to marginMatrix}
\usage{
addMargin(marginMatrix, varName, vecTotals, adjustToOne = TRUE,
  thresholdAdjustToOne = 0.01)
}
\arguments{
\item{marginMatrix}{The matrix of margins to add the new margin to}

\item{varName}{Name of variable in calibration matrix corresponding
to the new margin}

\item{vecTotals}{values of margins (Calmar style) for the variable.
Note : if length(vecTotals) > 1, then sum(thresholdAdjustToOne) has to be 1.}

\item{adjustToOne}{if TRUE and sum(vecTotals) is nearly 1, modify values of vecTotals
so that sum is 1.}

\item{thresholdAdjustToOne}{adjust sum(vecTotals) to 1 if difference
is under thresholdAdjustToOne}
}
\description{
Adds a margin to marginMatrix
}

