% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fonctions_grh.R
\name{ajouterPoidsGRH}
\alias{ajouterPoidsGRH}
\title{Create Non-response Adjusted Weights using the HRG method}
\usage{
ajouterPoidsGRH(data, modelGRH, colPoids = "POIDS", colNew = c("pHat",
  "numGRH", "rapportPoidsCNR", "POIDS_CNR"), nGRH = NULL, stats = FALSE)
}
\arguments{
\item{data}{The dataframe containing the survey results. Must include
every variable needed by modelGRH as well as a column containing the sampling
weights for each responding unit}

\item{modelGRH}{The logit model predicting the non-response}

\item{colPoids}{The column name of the column containing the weights
in the dataframe}

\item{colNew}{Specifies the name of the columns added to the dataframe}

\item{nGRH}{Number of groups created. If NULL, the number of groups is computed
using the Beaumont-Haziza method}

\item{If}{TRUE, display a few stats and graphs about the NRA weights created}
}
\description{
Create Non-response Adjusted Weights using the HRG method
}

