% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calmarFunctions.R
\name{calibrationMarginStats}
\alias{calibrationMarginStats}
\title{Gives stats about the calibration process: totals after/before calibration vs. margins
Same as first panels output in Calmar/Calmar 2}
\usage{
calibrationMarginStats(data, marginMatrix, popTotal = NULL,
  colWeights = "POIDS", colCalibratedWeights = NULL, calibThreshold = 1)
}
\arguments{
\item{data}{dataframe containing the survey data}

\item{marginMatrix}{matrix of margins}

\item{popTotal}{total of population, useful if margins are entered in relative value}

\item{colWeights}{name of weights column in the dataframe}
}
\description{
Gives stats about the calibration process: totals after/before calibration vs. margins
Same as first panels output in Calmar/Calmar 2
}

