% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmarFunctions.R
\name{checkMarginMatrix}
\alias{checkMarginMatrix}
\title{Check validity of marginMatrix (deprecated)}
\usage{
checkMarginMatrix(marginMatrix)
}
\description{
Check validity of marginMatrix (deprecated)
}

