% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmarFunctions.R
\name{checkNumberMargins}
\alias{checkNumberMargins}
\title{Checks if number of modalities in data matches expected ones according
to marginMatrix}
\usage{
checkNumberMargins(data, marginMatrix)
}
\description{
Checks if number of modalities in data matches expected ones according
to marginMatrix
}

