% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modalities_functions.R
\name{colToDummies}
\alias{colToDummies}
\title{Changes a column containing multiple values
to a matrix of columns containing the dummies
corresponding to each value.}
\usage{
colToDummies(col, nameCol, modalities = NULL, keepValue = FALSE)
}
\arguments{
\item{col}{input column}

\item{nameCol}{name that will be used as a prefix for
dummies column name in the output matrix}

\item{modalities}{if a vector is entered, dummies produced
will only be the ones corresponding to the values in the
"modalities" input column + another one containing all
the other modalities.}

\item{keepValue}{Logical. If TRUE, puts not "1"s in the dummies output
columns but the real values in the "col" column (except if values are
non-numeric)}
}
\value{
Matrix containing the dummy columns
}
\description{
Changes a column containing multiple values
to a matrix of columns containing the dummies
corresponding to each value.
}
