% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Count.Matrix}
\alias{Count.Matrix}
\title{Count.Matrix}
\format{
A data frame with 57,133 rows and 44 variables.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE197067}
}
\usage{
Count.Matrix
}
\description{
Raw Count Matrix of RNA-Seq samples from published study Rade M, Böhlen S, Neuhaus V, Löffler D et al. A time-resolved meta-analysis of consensus gene expression profiles during human T-cell activation. Genome Biol 2023 Dec 14;24(1):287. PMID: 38098113.
}
\keyword{datasets}
