% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmarFunctions.R
\name{createCalibrationMatrix}
\alias{createCalibrationMatrix}
\title{ensures compatibility with first version of icarus 
(when it was still called gaston 0.0.1)}
\usage{
createCalibrationMatrix(marginMatrix, data, popVector = TRUE)
}
\description{
ensures compatibility with first version of icarus 
(when it was still called gaston 0.0.1)
}

