% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data}
\alias{data}
\title{A small example sample to compare tolerance definition between Icarus and Calmar2 (SAS)}
\usage{
data
}
\description{
This table features a samples of 100000 units. 
This table contains : 
\itemize{\item 25 variables named \code{varn} where n denotes the number of the variables.
\item \code{w_HT} a variable corresponding to the initial weight (associated to the Horvitz-Thompson estimator).
\item \code{w_cale_icarus} a variable corresponding to calibrated weight computed with icarus.}
}
\author{
Khaled Larbi
}
