% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{data_ex2}
\alias{data_ex2}
\title{A small example sample for calibration with Icarus}
\format{15 rows, one per unit in sample.
1 column "id", unique id for each unit.
4 columns of auxiliary variables ("service", "categ", "sexe", "salaire").
1 column "cinema" - the variable of interest
1 column "weight" - the Horvitz-Thompson weights}
\usage{
data_ex2
}
\description{
This table features a samples of 15 units (drawn from a population of size 300),
used in a small survey to determine how frequently the employees of a firm go the movies (column "cinema").
Some auxiliary variables are given, which allows the use of calibration to improve
estimates. Margins for these auxiliary variables are known:
categ: 80 (modality 1) ; 90 (modality 2) ; 60 (modality 3)
sexe: 140 (modality 1) ; 90 (modality 2)
service: 100 (modality 1) ; 130 (modality 2)
salaire : 470000
}
\author{
Antoine Rebecq
}

