% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus_utils.R
\name{Direction_correction}
\alias{Direction_correction}
\title{Direction_correction}
\usage{
Direction_correction(Matrix, Group)
}
\description{
This function is used to correct the direction of the signature from different runs.
}
