% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus_utils.R
\name{faster_ICA}
\alias{faster_ICA}
\title{faster_ICA}
\usage{
faster_ICA(
  whitening_list,
  n.comp,
  alg.typ = c("parallel", "deflation"),
  fun = c("logcosh", "exp"),
  row.norm = FALSE,
  maxit = 200,
  tol = 1e-04,
  alpha = 1,
  w.init = NULL,
  method = "R"
)
}
\description{
A faster implementation of fastICA from fastICA package
}
