% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus_utils.R
\name{fasterICA_def}
\alias{fasterICA_def}
\title{fasterICA_def}
\usage{
fasterICA_def(X, n.comp, tol, fun, alpha, maxit, w.init)
}
\description{
A faster implementation of fastICA_def from fastICA package
}
