% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus_utils.R
\name{fasterICA_par}
\alias{fasterICA_par}
\title{fasterICA_par}
\usage{
fasterICA_par(X, n.comp, tol, fun, alpha, maxit, verbose, w.init)
}
\description{
A faster implementation of fastICA_par from fastICA package
}
