% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus_utils.R
\name{find_kneedle_point}
\alias{find_kneedle_point}
\title{find_kneedle_point}
\usage{
find_kneedle_point(x, y)
}
\description{
This function is used to correct the direction of the signature from different runs.
}
