% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{HTtotal}
\alias{HTtotal}
\title{Weighted estimator for total}
\usage{
HTtotal(var, weights)
}
\arguments{
\item{var}{column of variable of interest}

\item{weights}{column of weights matching the variable of interest}
}
\value{
Estimated total
}
\description{
Computes the weighted estimator for the total of a column. Alias for
\code{\link{weightedTotal}}
}
\seealso{
\code{\link{weightedTotal}}
}

