% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmarFunctions.R
\name{HTtotals}
\alias{HTtotals}
\title{private function that computes weighted estimates}
\usage{
HTtotals(dummyModalitiesMatrix, weights)
}
\description{
private function that computes weighted estimates
}

