% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus.R
\name{ICARus}
\alias{ICARus}
\title{ICARus}
\usage{
ICARus(
  Matrix,
  numberofcomponents,
  iteration = 100,
  numberofcores = 2,
  distance_measure = c("pearson", "euclidean"),
  clustering_algorithm = c("Hierarchical"),
  Hierarchical.clustering.method = c("ward.D2", "ward.D", "single", "single", "complete",
    "average", "mcquitty", "median", "centroid"),
  scale = T,
  ...
)
}
\arguments{
\item{Matrix}{A Matrix where rows are features and columns are observations.}

\item{numberofcomponents}{Number of independent component to compute, must be an integer larger than 1.}

\item{iteration}{The number of iterations of ICA to be run, default is 100.}

\item{numberofcores}{Number of cores to use. The default is 2.}

\item{distance_measure}{Choose which distance measurement to use for signatures Currently the pipeline supports pearson correlation coefficient ('pearson') and euclidean distance ('euclidean'). The default is 'pearson'.}

\item{clustering_algorithm}{Choose which clustering algorithm to use. Currently the pipeline supports Hierarchical clustering ('Hierarchical') and a clustering method based on Gale Shapely Algorithm ('MatchMaking'). The default is 'Hierarchical'.}

\item{Hierarchical.clustering.method}{Choose which hierarchical clustering algorithm to use. The default is 'ward.D2'.}
}
\value{
Three Matrix: 1. Stability of independent components. 2. The "A" matrix from ICA. 3. The "S" matrix from ICA.
}
\description{
This function allows parallel running of multiple Independent component analysis (ICA) at the same time,
which can be used to evaluate the robustness and stability of the independent components being extracted.
}
