% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus.R
\name{ICARus_complete}
\alias{ICARus_complete}
\title{ICARus_complete}
\usage{
ICARus_complete(
  Matrix,
  measure = c("cumulative_proportion", "standard_deviation"),
  iteration = 100,
  numberofcores = 4,
  numbers_of_parameter_for_reproducibility_test = 10,
  distance_measure = c("pearson", "euclidean"),
  clustering_algorithm = c("Hierarchical"),
  Hierarchical.clustering.method = c("ward.D2", "ward.D", "single", "single", "complete",
    "average", "mcquitty", "median", "centroid"),
  scale = T,
  tolerance = 1e-06,
  max.iteration = 1000,
  upperbound = 100,
  quality.index.threshold = 0.75,
  lowerbound = 50
)
}
\arguments{
\item{Matrix}{A Matrix where rows are features and columns are observations. The count matrix needs to be normalized and log transformed before being input.}

\item{measure}{How to estimate the number of independent components. Default is cumulative_proportion.}

\item{iteration}{The number of iterations of ICA to be run, default is 100.}

\item{numberofcores}{Number of cores to use. The default is 2.}

\item{distance_measure}{Choose which distance measurement to use for signatures Currently the pipeline supports pearson correlation coefficient ('pearson') and euclidean distance ('euclidean'). The default is 'pearson'.}

\item{clustering_algorithm}{Choose which clustering algorithm to use. Currently the pipeline supports Hierarchical clustering ('Hierarchical') and a clustering method based on Gale Shapely Algorithm ('MatchMaking'). The default is 'Hierarchical'.}

\item{Hierarchical.clustering.method}{Choose which hierarchical clustering algorithm to use. The default is 'ward.D2'.}

\item{tolerance}{The tolerance input for fastICA to converge, the default is 1e-10.}

\item{max.iteration}{The maximum number of iteration of fastICA. Default is 10000.}

\item{upperbound}{For each parameter x the fastICA is performed y times specified by user. Then the clustering algorithm will cluster the resultsinto x clusters. Ideally each cluster will have the size of y, but that is not always the case. This parameter sets the upper bound for the size of the cluster for the signature to be considered 'stable'.}

\item{quality.index.threshold}{For each parameter, the signatures obtained by fastICA are evaluated using the index proposed by icasso. The index ranges from 0 to 1, where 0 means unstable at all and 1 means perfectly stable. The threshold specify the lower bound of index for each signature.}

\item{lowerbound}{For each parameter x the fastICA is performed y times specified by user. Then the clustering algorithm will cluster the resultsinto x clusters. Ideally each cluster will have the size of y, but that is not always the case. This parameter sets the lower bound for the size of the cluster for the signature to be considered 'stable'.}
}
\value{
Three Matrix: 1. Stability of independent components. 2. The "A" matrix from ICA. 3. The "S" matrix from ICA.
}
\description{
This function is the main function of the package ICARus. The function automatically estimates the number of independent components to extract and output the most stable ones.
}
