% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus.R
\name{ICARus_est}
\alias{ICARus_est}
\title{ICARus_est}
\usage{
ICARus_est(
  Matrix,
  parameter_set,
  iteration = 100,
  numberofcores = 1,
  clustering_algorithm = "complete",
  ...
)
}
\arguments{
\item{Matrix}{A Matrix where rows are features and columns are observations.}

\item{parameter_set}{A vector of integers. Every one of the integer in the vector needs to be larger than 2.}

\item{iteration}{The number of iterations of ICA to be run, default is 100.}

\item{numberofcores}{Number of threads to use. The default is 1}

\item{clustering_algorithm}{The hierarchical clustering algorithm for clustering the independent component analysis. Default is 'complete'.}
}
\value{
Three Matrix: 1. Stability of independent components. 2. The "A" matrix from ICA. 3. The "S" matrix from ICA.
}
\description{
This function can be used to evaluate different parameters for ICA. In the parameter "parameter_set" user can input multiple integers and compare which parameter is a better one.
}
