% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus.R
\name{ICARus_ultrafast}
\alias{ICARus_ultrafast}
\title{ICARus_ultrafast}
\usage{
ICARus_ultrafast(
  Matrix,
  numberofcores = 4,
  numbers_of_parameter_for_reproducibility_test = 10,
  distance_measure = c("pearson", "euclidean"),
  clustering_algorithm = c("Hierarchical"),
  Hierarchical.clustering.method = c("ward.D2", "ward.D", "single", "single", "complete",
    "average", "mcquitty", "median", "centroid"),
  tolerance = 1e-10,
  max.iteration = 50000
)
}
\arguments{
\item{Matrix}{A Matrix where rows are features and columns are observations. The count matrix needs to be normalized and log transformed before being input.}

\item{numberofcores}{Number of cores to use. The default is 2.}

\item{distance_measure}{Choose which distance measurement to use for signatures Currently the pipeline supports pearson correlation coefficient ('pearson') and euclidean distance ('euclidean'). The default is 'pearson'.}

\item{clustering_algorithm}{Choose which clustering algorithm to use. Currently the pipeline supports Hierarchical clustering ('Hierarchical') and a clustering method based on Gale Shapely Algorithm ('MatchMaking'). The default is 'Hierarchical'.}

\item{Hierarchical.clustering.method}{Choose which hierarchical clustering algorithm to use. The default is 'ward.D2'.}

\item{tolerance}{The tolerance input for fastICA to converge, the default is 1e-10.}

\item{max.iteration}{The maximum number of iteration of fastICA. Default is 10000.}
}
\value{
Three Matrix: 1. Stability of independent components. 2. The "A" matrix from ICA. 3. The "S" matrix from ICA.
}
\description{
This function is the main function of the package ICARus. The function automatically estimates the number of independent components to extract and output the most stable ones.
}
