% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus_utils.R
\name{Individual_Clustering}
\alias{Individual_Clustering}
\title{Individual_Clustering}
\usage{
Individual_Clustering(Matrix, Group, ncluster, method = "complete")
}
\description{
This function cluster the ICA results from iterations
}
