% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus_utils.R
\name{Individual_Matching}
\alias{Individual_Matching}
\title{Individual_Matching}
\usage{
Individual_Matching(Matrix, Group, ncluster)
}
\description{
This function matches signatures from different iterations using galeShapley.marriageMarket
}
