% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus_utils.R
\name{IR_Calculation}
\alias{IR_Calculation}
\title{IR_Calculation}
\usage{
IR_Calculation(Correlation_Matrix, Clustering_identity, numberofcores = 6)
}
\description{
This function generates the IR index for ICA
}
