% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus_utils.R
\name{kneedle}
\alias{kneedle}
\title{kneedle}
\usage{
kneedle(x, y, decreasing, concave, sensitivity = 1)
}
\arguments{
\item{x}{A vector of x coordinates.}

\item{y}{A vector of y coordinates.}

\item{sensitivity}{How sensitive should knee detection be? Defaults to 1. Can be a float/double.}
}
\value{
This function returns a x,y coordinate pair that corresponds to the knee point
}
\description{
The kneedle algorithm implemented by Emmet Tam. Pulled directly from https://github.com/etam4260/kneedle.
}
