% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{margins}
\alias{margins}
\title{A small example sample to compare tolerance definition between Icarus and Calmar2 (SAS).}
\format{
\itemize{\item \code{name} : the name of the variables
\item \code{int} : 0 because all variables are quantitatives in our example
\item marges : the margin associated to the variable}
}
\usage{
margins
}
\description{
This table contains all margins associated to the 25 variables from \code{data}
}
\author{
Khaled Larbi
}
