% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{metadata}
\format{
A data frame with 44 rows and 3 variables:
\describe{
\item{Replicates}{The replicate identities of the sample.}
\item{Timepoint}{The time when the sample is obtained.}
\item{Condition}{The stimulation condition of the sample.}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE197067}
}
\usage{
metadata
}
\description{
Metadata of RNA-Seq samples from published study Rade M, Böhlen S, Neuhaus V, Löffler D et al. A time-resolved meta-analysis of consensus gene expression profiles during human T-cell activation. Genome Biol 2023 Dec 14;24(1):287. PMID: 38098113.
}
\keyword{datasets}
