% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmarFunctions.R
\name{missingValuesMargins}
\alias{missingValuesMargins}
\title{Displays number of NAs among margins}
\usage{
missingValuesMargins(data, marginMatrix)
}
\description{
Displays number of NAs among margins
}

