% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fonctions_grh.R
\name{ngrhHazizaBeaumont}
\alias{ngrhHazizaBeaumont}
\title{Computes the optimal number of HRG according to the Beaumont-Haziza method}
\usage{
ngrhHazizaBeaumont(nGRHtests, data, seuil = 0.99, colpHat = "pHat")
}
\description{
Computes the optimal number of HRG according to the Beaumont-Haziza method
}

