% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus.R
\name{PCA.Estimation}
\alias{PCA.Estimation}
\title{PCA.Estimation}
\usage{
PCA.Estimation(
  Matrix = NULL,
  measure = c("standard_deviation", "cumulative_proportion")
)
}
\arguments{
\item{Matrix}{A Matrix where rows are features and columns are observations.}
}
\value{
A list of one vector and one ggplot,
}
\description{
This function performs PCA analysis on a given matrix, and estimate the best number of independent components by finding the elbow point in the variance explained elbow plot.
}
