% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{poptest_calmar}
\alias{poptest_calmar}
\title{Calibration on population test - made on Calmar2}
\format{
1000 rows, one per unit in the sample.
1 column "ident", with a unique id for every unit in the sample
3 methods of calibration are used (linear, raking, and logit with bounds LO=0.2 and UP=1.3) for
two different margins tables \code{\link{table_margins_1}} and \code{\link{table_margins_2}}, which results in 
7 columns of weights.
}
\usage{
poptest_calmar
}
\description{
This data set features calibration weights for the sample test of \code{\link{dataPop}}
(using margins tables \code{\link{table_margins_1}} and \code{\link{table_margins_2}}). Calibration is
is computed using the SAS Macro Calmar2, for test purposes.
}
\references{
Le Guennec, J., and Sautory, O. (2002). Calmar 2: Une nouvelle version 
de la macro calmar de redressement d'echantillon par calage. 
Journees de Methodologie Statistique, Paris. INSEE.
}
\author{
Antoine Rebecq
}
