% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calmarFunctions.R
\name{regroupCalibrationModalities}
\alias{regroupCalibrationModalities}
\title{Regroups modalities entered in "vecModalities" into single
"newModality" in "calibrationMatrix" and adapts "marginMatrix" to the new concept.
Typical usage is right before a calibration (and after comptutation of marginMatrix), when
you realise calibration output is better when several modalities are reduced to one.
(typically very rare modalities, on which calibration constraints are very restrictive).
Uses pseudo-"call by reference" via eval.parent because 2 objects are modified :
calibrationMatrix and marginMatrix}
\usage{
regroupCalibrationModalities(calibrationMatrix, marginMatrix,
  calibrationVariable, vecModalities, newModality)
}
\description{
Regroups modalities entered in "vecModalities" into single
"newModality" in "calibrationMatrix" and adapts "marginMatrix" to the new concept.
Typical usage is right before a calibration (and after comptutation of marginMatrix), when
you realise calibration output is better when several modalities are reduced to one.
(typically very rare modalities, on which calibration constraints are very restrictive).
Uses pseudo-"call by reference" via eval.parent because 2 objects are modified :
calibrationMatrix and marginMatrix
}

