% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{results_from_sas}
\alias{results_from_sas}
\title{A small example sample to compare tolerance definition between Icarus and Calmar2 (SAS).}
\usage{
margins
}
\description{
This table contains calibrated weights computed using SAS (Calmar) for different calibration tolerances.
}
\author{
Khaled Larbi
}
