% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RugCreate.R
\name{RugCreate}
\alias{RugCreate}
\title{Creates "RUG" plots for binomial GLMs}
\usage{
RugCreate(x, y, num.bins)
}
\arguments{
\item{x:}{Your independent variable.  Only GLMs with 1 independent variable are supported}

\item{y:}{Your dependent variable.}

\item{num.bins:}{The number of bins to break your data into}
}
\description{
This function creates "RUG" plots that can be used to test for overfitting of binomial GLMs
}
\examples{
RugCreate()
}
\keyword{Binomial}
\keyword{GLM}
\keyword{Icarus}

