% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/seq_matrix.R
\name{seq.matrix}
\alias{seq.matrix}
\title{A Sequentially Numbered Matrix Creating Function}
\usage{
\method{seq}{matrix}(number, name, nrow, ncol)
}
\arguments{
\item{number:}{The number of matrices you wish to create}

\item{name:}{The common name of the matrices}

\item{nrow:}{The number of rows you want to be present in the matrices you create}

\item{ncol:}{The number of columns you want to be present in teh matrices you create}
}
\description{
This function creates matrices of identical size and name, with a numeric identifier
}
\examples{
seq.matrix()
}
\keyword{Icarus}
\keyword{Matrix}
\keyword{Multiple}

