% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus.R
\name{Signature_Hierarchical_Clustering}
\alias{Signature_Hierarchical_Clustering}
\title{PCA.Estimation}
\usage{
Signature_Hierarchical_Clustering(
  Disimmilarity,
  Affiliation.Matrix,
  Signature.Matrix,
  min_cluster = 2,
  max_cluster = ncol(Disimmilarity)/2,
  numberofcores = 2,
  ...
)
}
\arguments{
\item{Disimmilarity}{A dissimilarity matrix.}

\item{Affiliation.Matrix}{The Affiliation Matrix output by ICARus.}

\item{Signature.Matrix}{The Signature Matrix output by ICARus.}

\item{min_cluster}{Minimum number of clusters to be tested, default is 2.}

\item{max_cluster}{Maximum number of clusters to be tested, default is half of number of columns in  Signature Matrix.}

\item{numberofcores}{Number of cores to use. The default is 2.}
}
\value{
A list of one vector and one ggplot,
}
\description{
This function estimate the optimal number of clusters in a group of independent component analysis results.
}
