% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fonctions_grh.R
\name{statsMarges}
\alias{statsMarges}
\title{Outputs stats about the non-response adjustment process for some variables}
\usage{
statsMarges(data, vecMarges, colPoids = "POIDS", colPoidsCNR = "POIDS_CNR",
  colRepondant = "repondant", sepDummies = "_", statsEchantillon = FALSE)
}
\arguments{
\item{vecMarges}{names of variables for which stats are computed}
}
\value{
List of stats for each vector
}
\description{
Outputs stats about the non-response adjustment process for some variables
}

