% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fonctions_grh.R
\name{statsTauxCollecte}
\alias{statsTauxCollecte}
\title{Outputs stats about the collection rate}
\usage{
statsTauxCollecte(data, nameCol, nameDummy = nameCol,
  colRepondant = "repondant", selection = c("auto"), sepDummies = "_")
}
\description{
Outputs stats about the collection rate
}

