% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{table_margins_1}
\alias{table_margins_1}
\title{Margins for calibration of test population}
\format{
A margins table written in the Icarus format.
}
\usage{
table_margins_1
}
\description{
This table features calibration margins for the sample of the test population
of \code{\link{dataPop}}
}
\author{
Antoine Rebecq
}
