% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimators.R
\name{HTmean}
\alias{HTmean}
\title{Weighted estimator for the mean}
\usage{
HTmean(var, weights, popTot = NULL)
}
\arguments{
\item{var}{column of variable of interest}

\item{weights}{column of weights matching the variable of interest}
}
\value{
Estimated mean
}
\description{
Computes the weighted estimator for the mean of a column. Alias for
\code{\link{weightedMean}}
}
\seealso{
\code{\link{weightedMean}}
}

