% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus.R
\name{ICARus}
\alias{ICARus}
\title{ICARus}
\usage{
ICARus(
  Matrix,
  numberofcomponents,
  iteration = 100,
  numberofcores = 2,
  clustering_algorithm = "complete",
  ...
)
}
\arguments{
\item{Matrix}{A Matrix where rows are features and columns are observations.}

\item{numberofcomponents}{Number of independent component to compute, must be an integer larger than 1.}

\item{iteration}{The number of iterations of ICA to be run, default is 100.}

\item{numberofcores}{Number of threads to use. The default is 2.}

\item{clustering_algorithm}{The hierarchical clustering algorithm for clustering the independent component analysis. Default is 'complete'.}
}
\value{
Three Matrix: 1. Stability of independent components. 2. The "A" matrix from ICA. 3. The "S" matrix from ICA.
}
\description{
This function allows parallel running of multiple Independent component analysis (ICA) at the same time,
which can be used to evaluate the robustness and stability of the independent components being extracted.
}
