% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus.R
\name{ICARus_est}
\alias{ICARus_est}
\title{ICARus_est}
\usage{
ICARus_est(
  Matrix,
  parameter_set,
  iteration = 100,
  numberofcores = 2,
  distance_measure = c("euclidean", "pearson"),
  clustering_algorithm = c("Hierarchical"),
  Hierarchical.clustering.method = c("ward.D2", "ward.D", "single", "single", "complete",
    "average", "mcquitty", "median", "centroid"),
  ...
)
}
\arguments{
\item{Matrix}{A Matrix where rows are features and columns are observations.}

\item{parameter_set}{A vector of integers. Every one of the integer in the vector needs to be larger than 2.}

\item{iteration}{The number of iterations of ICA to be run, default is 100.}

\item{numberofcores}{Number of cores to use. The default is 2.}

\item{distance_measure}{Choose which distance measurement to use for signatures Currently the pipeline supports pearson correlation coefficient ('pearson') and euclidean distance ('euclidean'). The default is 'pearson'.}

\item{clustering_algorithm}{Choose which clustering algorithm to use. Currently the pipeline supports Hierarchical clustering ('Hierarchical') and a clustering method based on Gale Shapely Algorithm ('MatchMaking'). The default is 'Hierarchical'.}

\item{Hierarchical.clustering.method}{Choose which hierarchical clustering algorithm to use. The default is 'ward.D2'.}
}
\value{
Three Matrix: 1. Stability of independent components. 2. The "A" matrix from ICA. 3. The "S" matrix from ICA.
}
\description{
This function can be used to evaluate different parameters for ICA. In the parameter "parameter_set" user can input multiple integers and compare which parameter is a better one.
}
