% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICARus_utils.R
\name{ParaICA}
\alias{ParaICA}
\title{ParaICA}
\usage{
ParaICA(
  CountMatrix,
  faster_whiten = NULL,
  numberofcomponents,
  iteration,
  numberofcores = 4,
  ...
)
}
\description{
This function runs ICA using multiple threads.
}
