% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modalities_functions.R
\name{regroupModalities}
\alias{regroupModalities}
\title{Regroup}
\usage{
regroupModalities(column, regroupMatrix, modalities = NULL)
}
\arguments{
\item{column}{Column vector which values are going to be replaced}

\item{regroupMatrix}{Bounds of the values to regroup under the same modality}

\item{modalities}{Specify the values of the modalities to use. Must match number of rows
of regroupMatrix
If not specified, replacement modalities will be 1:length(column)}
}
\value{
Column vector with regrouped modalities
}
\description{
Regroup
}

