% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimators.R
\name{weightedMean}
\alias{weightedMean}
\title{Weighted estimator for the mean}
\usage{
weightedMean(var, weights, popTot = NULL)
}
\arguments{
\item{var}{column of variable of interest}

\item{weights}{column of weights matching the variable of interest}
}
\value{
Estimated mean
}
\description{
Computes the weighted estimator for the mean of a column
}
\seealso{
\code{\link{HTmean}}
}

